requirements.txt
Flask_SocketIO==5.1.0
Flask_Mail==0.9.1
Flask_Migrate==3.0.1
WTForms==2.3.3
alembic==1.6.5
Flask_WTF==0.14.3
jsonpickle==2.0.0
Flask_SQLAlchemy==2.5.1
Flask==1.1.2
SQLAlchemy==1.4.15
Flask_Bcrypt==0.7.1
Flask_Login==0.5.0
Pyrebase4==4.5.0
Pillow==8.3.1
python-dotenv==0.19.0
gevent-websocket==0.10.1
gunicorn==19.9.0
psycopg2==2.9.1
eventlet==0.30.2
email-validator==1.1.3
PushySDK==0.1.5